src/requirements.txt
psycopg2-binary==2.7.5
django==2.0.6
json_log_formatter==0.2.0
unittest-xml-reporting==2.2.0
djangorestframework==3.8.2
coreapi==2.3.3
django-rest-swagger==2.2.0